Move `FileLock` into `oak_cache` and remove `oak_fs`
