--- crates/oak_fs/Cargo.toml
@@ -1,18 +1,0 @@
-[package]
-name = "oak_fs"
-version = "0.1.0"
-authors.workspace = true
-edition.workspace = true
-rust-version.workspace = true
-license.workspace = true
-
-[dependencies]
-anyhow.workspace = true
-libc.workspace = true
-log.workspace = true
-
-[target.'cfg(windows)'.dependencies]
-windows-sys = { workspace = true, features = ["Win32_Foundation"] }
-
-[lints]
-workspace = true
