--- crates/oak_fs/src/lib.rs
@@ -1,2 +1,0 @@
-pub mod file_lock;
-pub mod permissions;
