--- crates/oak_fs/src/permissions.rs
@@ -1,8 +1,0 @@
-use std::path::Path;
-
-/// Mark a file as read only
-pub fn set_readonly(path: &Path) -> std::io::Result<()> {
-    let mut permissions = std::fs::metadata(path)?.permissions();
-    permissions.set_readonly(true);
-    std::fs::set_permissions(path, permissions)
-}
