+++ crates/oak_source/Cargo.toml
@@ -11,7 +11,6 @@ anyhow.workspace = true
 flate2.workspace = true
 log.workspace = true
 oak_cache.workspace = true
-oak_fs.workspace = true
 tar.workspace = true
 ureq.workspace = true
