+++ crates/oak_srcref/Cargo.toml
@@ -11,7 +11,6 @@ anyhow.workspace = true
 hex.workspace = true
 log.workspace = true
 oak_cache.workspace = true
-oak_fs.workspace = true
 oak_r_process.workspace = true
 sha2.workspace = true
